added appearance settings for all notes

+++ app/src/main/java/com/furianrt/mydiary/data/DataManager.kt
@@ -105,7 +105,7 @@ interface DataManager {
 
     fun getTextColor(): Int
 
-    fun getTextSize(): Float
+    fun getTextSize(): Int
 
     fun getNoteBackgroundColor(): Int
